Bubble sort

+++ CSV/CsvObject.cs
@@ -10,6 +10,6 @@ namespace CSV
 {
     public class CsvObject
     {
-        public int Id { get; set; }
+        public double Id { get; set; }
     }
 }
